接続切断時にタイトルに戻るように修正

+++ NetworkGame/NetworkGame/Device/Input/Mouse.cpp
@@ -10,7 +10,7 @@ Input::KeyCode cast(Input::MouseButton button) {
 }
 
 Math::Vector2 toVector2(const POINT& pos) {
-    return Math::Vector2(pos.x, pos.y);
+    return Math::Vector2((float)pos.x, (float)pos.y);
 }
 
 POINT toPoint(const Math::Vector2& pos) {
